Remove v3 validation.
