+++ aquarius/ddo_checker/shacl_schemas/v4/remote_v4.0.0.ttl
@@ -21,18 +21,7 @@ schema:DDOShape
         sh:minCount 1 ;
         sh:maxCount 1 ;
     ] ;
-    sh:property [
-        sh:path schema:created ;
-        sh:datatype xsd:dateTime;
-        sh:minCount 1 ;
-        sh:maxCount 1 ;
-    ] ;
-    sh:property [
-        sh:path schema:updated;
-        sh:datatype xsd:dateTime ;
-        sh:minCount 1 ;
-        sh:maxCount 1 ;
-    ] ;
+    # TODO reinstate created and updated
     sh:property [
         sh:path schema:chainId;
         sh:datatype xsd:integer ;
